+++ app/Modules/Admin/routes/web.php
@@ -4,4 +4,4 @@ use Codemonster\Xen\Modules\Admin\Controllers\DashboardController;
 use Codemonster\Xen\Modules\Auth\Middleware\AuthMiddleware;
 
 router()->get('/admin', [DashboardController::class, 'index'])
-    ->middleware(AuthMiddleware::class, 'admin');
+    ->middleware(AuthMiddleware::class, 'admin|strict');