+++ bootstrap/providers/SecurityServiceProvider.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Codemonster\Xen\Providers;
+
+use Codemonster\Security\Providers\SecurityServiceProvider as BaseSecurityServiceProvider;
+
+class SecurityServiceProvider extends BaseSecurityServiceProvider {}